requirements-test.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-test.txt requirements-test.in
#
biopython==1.85
    # via -r requirements.txt
certifi==2024.2.2
    # via
    #   -r requirements.txt
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements.txt
    #   requests
ensembl-metadata-api @ git+https://github.com/Ensembl/ensembl-metadata-api.git@3.0.1
    # via -r requirements.txt
ensembl-py==2.1.3
    # via -r requirements.txt
ensembl-utils==0.5.0
    # via -r requirements.txt
exceptiongroup==1.2.2
    # via
    #   -r requirements.txt
    #   pytest
greenlet==3.1.1
    # via
    #   -r requirements.txt
    #   sqlalchemy
idna==3.6
    # via
    #   -r requirements.txt
    #   requests
iniconfig==2.0.0
    # via
    #   -r requirements.txt
    #   pytest
mock==5.2.0
    # via -r requirements-test.in
numpy==2.2.6
    # via
    #   -r requirements.txt
    #   biopython
packaging==24.1
    # via
    #   -r requirements.txt
    #   pytest
pluggy==1.5.0
    # via
    #   -r requirements.txt
    #   pytest
pydantic==1.10.14
    # via -r requirements.txt
pymysql==1.1.1
    # via -r requirements-test.in
pytest==8.3.3
    # via
    #   -r requirements.txt
    #   -r requirements-test.in
    #   ensembl-utils
    #   pytest-env
pytest-env==1.1.5
    # via -r requirements-test.in
python-dotenv==1.0.1
    # via
    #   -r requirements.txt
    #   ensembl-utils
pyyaml==6.0.2
    # via
    #   -r requirements.txt
    #   ensembl-utils
requests==2.31.0
    # via
    #   -r requirements.txt
    #   ensembl-utils
sqlalchemy==2.0.35
    # via
    #   -r requirements.txt
    #   ensembl-py
    #   ensembl-utils
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via
    #   -r requirements.txt
    #   ensembl-utils
tomli==2.0.1
    # via
    #   -r requirements.txt
    #   pytest
    #   pytest-env
typing-extensions==4.10.0
    # via
    #   -r requirements.txt
    #   pydantic
    #   sqlalchemy
unidecode==1.4.0
    # via -r requirements.txt
urllib3==1.26.18
    # via
    #   -r requirements.txt
    #   requests
wget==3.2
    # via -r requirements.txt
